+++ passkit/lib/src/crypto/rsa_key_parser.dart
@@ -1,0 +1,78 @@
+import 'dart:convert' as convert;
+import 'dart:typed_data';
+
+import 'package:pointycastle/pointycastle.dart';
+
+/// RSA PEM parser.
+/// Inspired by https://pub.dev/documentation/encrypt/latest/encrypt/RSAKeyParser-class.html
+/// but adapted to make use of pointycastle instad of asn1lib.
+class RSAKeyParser {
+  /// Parses the PEM key no matter it is public or private, it will figure it out.
+  RSAAsymmetricKey parse(String key) {
+    final rows = key.split(RegExp(r'\r\n?|\n'));
+    final header = rows.first;
+
+    if (header == '-----BEGIN RSA PUBLIC KEY-----') {
+      return _parsePublic(_parseSequence(rows));
+    }
+
+    if (header == '-----BEGIN PUBLIC KEY-----') {
+      return _parsePublic(_pkcs8PublicSequence(_parseSequence(rows)));
+    }
+
+    if (header == '-----BEGIN RSA PRIVATE KEY-----') {
+      return _parsePrivate(_parseSequence(rows));
+    }
+
+    if (header == '-----BEGIN PRIVATE KEY-----') {
+      return _parsePrivate(_pkcs8PrivateSequence(_parseSequence(rows)));
+    }
+
+    throw FormatException('Unable to parse key, invalid format.', header);
+  }
+
+  RSAAsymmetricKey _parsePublic(ASN1Sequence sequence) {
+    final modulus = (sequence.elements![0] as ASN1Integer).integer!;
+    final exponent = (sequence.elements![1] as ASN1Integer).integer!;
+
+    return RSAPublicKey(modulus, exponent);
+  }
+
+  RSAAsymmetricKey _parsePrivate(ASN1Sequence sequence) {
+    final modulus = (sequence.elements![1] as ASN1Integer).integer!;
+    final exponent = (sequence.elements![3] as ASN1Integer).integer!;
+    final p = (sequence.elements![4] as ASN1Integer).integer!;
+    final q = (sequence.elements![5] as ASN1Integer).integer!;
+
+    return RSAPrivateKey(modulus, exponent, p, q);
+  }
+
+  ASN1Sequence _parseSequence(List<String> rows) {
+    final keyText = rows
+        .skipWhile((row) => row.startsWith('-----BEGIN'))
+        .takeWhile((row) => !row.startsWith('-----END'))
+        .map((row) => row.trim())
+        .join('');
+
+    final keyBytes = Uint8List.fromList(convert.base64.decode(keyText));
+    final asn1Parser = ASN1Parser(keyBytes);
+
+    return asn1Parser.nextObject() as ASN1Sequence;
+  }
+
+  ASN1Sequence _pkcs8PublicSequence(ASN1Sequence sequence) {
+    final ASN1Object bitString = sequence.elements![1];
+    final bytes = bitString.valueBytes!.sublist(1);
+    final parser = ASN1Parser(Uint8List.fromList(bytes));
+
+    return parser.nextObject() as ASN1Sequence;
+  }
+
+  ASN1Sequence _pkcs8PrivateSequence(ASN1Sequence sequence) {
+    final ASN1Object bitString = sequence.elements![2];
+    final bytes = bitString.valueBytes;
+    final parser = ASN1Parser(bytes);
+
+    return parser.nextObject() as ASN1Sequence;
+  }
+}